+++ pyabc/predictor/predictor.py
@@ -372,7 +372,7 @@ class GPPredictor(SimplePredictor):
 
     def __init__(
         self,
-        kernel: Callable | skl_gp.kernels.Kernel = None,
+        kernel: Callable | skl_gp.kernels.Kernel | None = None,
         normalize_features: bool = True,
         normalize_labels: bool = True,
         joint: bool = True,
@@ -438,8 +438,8 @@ class GPKernelHandle:
 
     def __init__(
         self,
-        kernels: list[str] = None,
-        kernel_kwargs: list[dict] = None,
+        kernels: list[str] | None = None,
+        kernel_kwargs: list[dict] | None = None,
         ard: bool = True,
     ):
         """
@@ -508,7 +508,7 @@ class MLPPredictor(SimplePredictor):
         normalize_features: bool = True,
         normalize_labels: bool = True,
         joint: bool = True,
-        hidden_layer_sizes: tuple[int, ...] | Callable = None,
+        hidden_layer_sizes: tuple[int, ...] | Callable | None = None,
         log_pearson: bool = True,
         **kwargs,
     ):
@@ -684,7 +684,7 @@ class ModelSelectionPredictor(Predictor):
         split_method: str = TRAIN_TEST_SPLIT,
         n_splits: int = 5,
         test_size: float = 0.2,
-        f_score: Callable = None,
+        f_score: Callable | None = None,
     ):
         """
         Parameters
@@ -722,6 +722,8 @@ class ModelSelectionPredictor(Predictor):
 
         if f_score is None:
             self.f_score = root_mean_square_error
+        else:
+            self.f_score = f_score
 
         # holds the chosen predictor model
         self.chosen_one: Predictor | None = None